Refactor project pages and components
