+++ src/routes.tsx
@@ -1,10 +1,10 @@
 import Home from "./pages/home/Home";
-import Projects from "./pages/Projects";
+import Projects from "./pages/projects/Projects";
 import React from "react";
 import {createBrowserRouter, redirect} from "react-router-dom";
 import Root from "./components/Root";
 import ErrorPage from "./pages/error-page";
-import Project from "./pages/Project";
+import Project from "./pages/projects/Project";
 import {loader as projectLoader} from "./routes/projects"
 
 export const router = createBrowserRouter([
